moving end to end test and changing Polyfit to PolyFit
