+++ app/src/main/java/com/github/se/polyfit/ui/components/AppTitleTopBar.kt
@@ -18,7 +18,7 @@ import com.github.se.polyfit.ui.utils.OverviewTags
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AppTitle(
-    titleText: String = LocalContext.current.getString(R.string.app_name),
+    titleText: String = LocalContext.current.getString(R.string.app_name_maj),
     alignment: Alignment = Alignment.TopCenter
 ) {
   TopAppBar(
